move getKiotaVersion to utilities folder
